Move the router to the repots controllers
