+++ src/controllers/report.controller.ts
@@ -7,13 +7,14 @@ import db from "../utils/dbconnection.util"
 import { courseModuleSchema, courseModuleUpdateSchema } from "../validations/courseModule.validationa";
 import { translationSchema, translationUpdateSchema } from "../validations/translation.validations";
 import ValidationsHolder from "../validations/validationHolder";
+import { quiz_survey_response } from '../models/quiz_survey_response.model';
 import BaseController from "./base.controller";
 import { constents } from "../configs/constents.config";
 import { mentor_course_topic } from "../models/mentor_course_topic.model";
 import { internal, notFound } from "boom";
 import { speeches } from "../configs/speeches.config";
 import ReportService from "../services/report.service";
-import { Op } from "sequelize";
+import { Op, QueryTypes } from 'sequelize';
 import { user } from "../models/user.model";
 
 export default class ReportController extends BaseController {
@@ -29,7 +30,223 @@ export default class ReportController extends BaseController {
     protected initializeRoutes(): void {
         //example route to add 
         this.router.get(`${this.path}/allMentorReports`, this.getAllMentorReports.bind(this));
+        this.router.get(`${this.path}/regList`, this.getMentorRegList.bind(this));
+        this.router.get(this.path + "/preSurvey", this.mentorPreSurvey.bind(this));
+        this.router.get(this.path + "/courseComplete", this.courseComplete.bind(this));
+        this.router.get(this.path + "/courseInComplete", this.courseInComplete.bind(this));
+        this.router.get(this.path + "/notRegistered", this.notRegistered.bind(this));
         // super.initializeRoutes();
+    }
+
+    protected async getMentorRegList(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
+        try {
+            const { quiz_survey_id } = req.params
+            const { page, size, status } = req.query;
+            let condition = {}
+            // condition = status ? { status: { [Op.like]: `%${status}%` } } : null;
+            const { limit, offset } = this.getPagination(page, size);
+            const modelClass = await this.loadModel(this.model).catch(error => {
+                next(error)
+            });
+            const paramStatus: any = req.query.status;
+            let whereClauseStatusPart: any = {};
+            let whereClauseStatusPartLiteral = "1=1";
+            let addWhereClauseStatusPart = false
+            if (paramStatus && (paramStatus in constents.common_status_flags.list)) {
+                whereClauseStatusPart = { "status": paramStatus }
+                whereClauseStatusPartLiteral = `status = "${paramStatus}"`
+                addWhereClauseStatusPart = true;
+            }
+            const mentorsResult = await mentor.findAll({
+                attributes: [
+                    "full_name",
+                    "mobile",
+                    "created_by",
+                    "created_at",
+                    "updated_at",
+                    "updated_by"
+                ],
+                raw: true,
+                where: {
+                    [Op.and]: [
+                        whereClauseStatusPart,
+                        condition
+                    ]
+                },
+                include: [
+                    {
+                        model: organization,
+                        attributes: [
+                            "organization_code",
+                            "district",
+                            "organization_name"
+                        ]
+                    },
+                    {
+                        model: user,
+                        attributes: [
+                            "username",
+                            "user_id"
+                        ]
+                    }
+                ],
+                limit, offset
+            });
+            if (!mentorsResult) {
+                throw notFound(speeches.DATA_NOT_FOUND)
+            }
+            if (mentorsResult instanceof Error) {
+                throw mentorsResult
+            }
+            res.status(200).send(dispatcher(res, mentorsResult, "success"))
+        } catch (err) {
+            next(err)
+        }
+    }
+    protected async mentorPreSurvey(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
+        try {
+            const { quiz_survey_id } = req.params
+            const { page, size, role } = req.query;
+            let condition = role ? role : 'MENTOR';
+            // let condition = role ? { role: { [Op.eq]: role } } : null;
+            const { limit, offset } = this.getPagination(page, size);
+            const modelClass = await this.loadModel(this.model).catch(error => {
+                next(error)
+            });
+            const paramStatus: any = req.query.status;
+            let whereClauseStatusPart: any = {};
+            let whereClauseStatusPartLiteral = "1=1";
+            let addWhereClauseStatusPart = false
+            if (paramStatus && (paramStatus in constents.common_status_flags.list)) {
+                whereClauseStatusPart = { "status": paramStatus }
+                whereClauseStatusPartLiteral = `status = "${paramStatus}"`
+                addWhereClauseStatusPart = true;
+            }
+            const mentorsResult = await quiz_survey_response.findAll({
+                attributes: [
+                    "quiz_response_id",
+                    "updated_at"
+                ],
+                raw: true,
+                where: {
+                    [Op.and]: [
+                        whereClauseStatusPart
+                    ]
+                },
+                include: [
+                    {
+                        model: user,
+                        attributes: [
+                            "full_name",
+                            "created_at",
+                            "updated_at"
+                        ],
+                        where: { role: condition }
+                    }
+                ],
+                limit, offset
+            });
+            if (!mentorsResult) {
+                throw notFound(speeches.DATA_NOT_FOUND)
+            }
+            if (mentorsResult instanceof Error) {
+                throw mentorsResult
+            }
+            res.status(200).send(dispatcher(res, mentorsResult, "success"))
+        } catch (err) {
+            next(err)
+        }
+    }
+    protected async courseComplete(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
+        try {
+            const { quiz_survey_id } = req.params
+            const { page, size, role } = req.query;
+            let condition = role ? { role: { [Op.eq]: role } } : null;
+            const { limit, offset } = this.getPagination(page, size);
+            const modelClass = await this.loadModel(this.model).catch(error => {
+                next(error)
+            });
+            const paramStatus: any = req.query.status;
+            let whereClauseStatusPart: any = {};
+            let whereClauseStatusPartLiteral = "1=1";
+            let addWhereClauseStatusPart = false
+            if (paramStatus && (paramStatus in constents.common_status_flags.list)) {
+                whereClauseStatusPart = { "status": paramStatus }
+                whereClauseStatusPartLiteral = `status = "${paramStatus}"`
+                addWhereClauseStatusPart = true;
+            }
+            const mentorsResult = await db.query("SELECT mentors.organization_code, mentors.district, mentors.full_name,(SELECT COUNT(mentor_topic_progress_id)FROM mentor_topic_progress AS mentor_progress WHERE mentor_progress.user_id=mentors.user_id) AS 'count' FROM mentors LEFT OUTER JOIN mentor_topic_progress AS mentor_progress ON mentors.user_id=mentor_progress.user_id where (SELECT COUNT(mentor_topic_progress_id)FROM mentor_topic_progress AS mentor_progress WHERE mentor_progress.user_id=mentors.user_id)= 9 GROUP BY mentor_progress.user_id", { type: QueryTypes.SELECT });
+            if (!mentorsResult) {
+                throw notFound(speeches.DATA_NOT_FOUND)
+            }
+            if (mentorsResult instanceof Error) {
+                throw mentorsResult
+            }
+            res.status(200).send(dispatcher(res, mentorsResult, "success"))
+        } catch (err) {
+            next(err)
+        }
+    }
+    protected async courseInComplete(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
+        try {
+            const { quiz_survey_id } = req.params
+            const { page, size, role } = req.query;
+            let condition = role ? { role: { [Op.eq]: role } } : null;
+            const { limit, offset } = this.getPagination(page, size);
+            const modelClass = await this.loadModel(this.model).catch(error => {
+                next(error)
+            });
+            const paramStatus: any = req.query.status;
+            let whereClauseStatusPart: any = {};
+            let whereClauseStatusPartLiteral = "1=1";
+            let addWhereClauseStatusPart = false
+            if (paramStatus && (paramStatus in constents.common_status_flags.list)) {
+                whereClauseStatusPart = { "status": paramStatus }
+                whereClauseStatusPartLiteral = `status = "${paramStatus}"`
+                addWhereClauseStatusPart = true;
+            }
+            const mentorsResult = await db.query("SELECT mentors.organization_code, mentors.district, mentors.full_name,(SELECT COUNT(mentor_topic_progress_id)FROM mentor_topic_progress AS mentor_progress WHERE mentor_progress.user_id=mentors.user_id) AS 'count' FROM mentors LEFT OUTER JOIN mentor_topic_progress AS mentor_progress ON mentors.user_id=mentor_progress.user_id where (SELECT COUNT(mentor_topic_progress_id)FROM mentor_topic_progress AS mentor_progress WHERE mentor_progress.user_id=mentors.user_id) != 9 GROUP BY mentor_progress.user_id", { type: QueryTypes.SELECT });
+            console.log(mentorsResult);
+            if (!mentorsResult) {
+                throw notFound(speeches.DATA_NOT_FOUND)
+            }
+            if (mentorsResult instanceof Error) {
+                throw mentorsResult
+            }
+            res.status(200).send(dispatcher(res, mentorsResult, "success"))
+        } catch (err) {
+            next(err)
+        }
+    }
+    protected async notRegistered(req: Request, res: Response, next: NextFunction): Promise<Response | void> {
+        try {
+            const { quiz_survey_id } = req.params
+            const { page, size, role } = req.query;
+            let condition = role ? { role: { [Op.eq]: role } } : null;
+            const { limit, offset } = this.getPagination(page, size);
+            const modelClass = await this.loadModel(this.model).catch(error => {
+                next(error)
+            });
+            const paramStatus: any = req.query.status;
+            let whereClauseStatusPart: any = {};
+            let whereClauseStatusPartLiteral = "1=1";
+            let addWhereClauseStatusPart = false
+            if (paramStatus && (paramStatus in constents.common_status_flags.list)) {
+                whereClauseStatusPart = { "status": paramStatus }
+                whereClauseStatusPartLiteral = `status = "${paramStatus}"`
+                addWhereClauseStatusPart = true;
+            }
+            const mentorsResult = await db.query("SELECT * FROM organizations WHERE NOT EXISTS(SELECT mentors.organization_code  from mentors WHERE organizations.organization_code = mentors.organization_code) ", { type: QueryTypes.SELECT });
+            if (!mentorsResult) {
+                throw notFound(speeches.DATA_NOT_FOUND)
+            }
+            if (mentorsResult instanceof Error) {
+                throw mentorsResult
+            }
+            res.status(200).send(dispatcher(res, mentorsResult, "success"))
+        } catch (err) {
+            next(err)
+        }
     }
 
     protected async getAllMentorReports(req:Request,res:Response,next:NextFunction){
